chore(ui): remove CardAnimation context
